requirements.txt
fastapi
python-jose[cryptography]
fastapi-sso
uvicorn
